{"types":"C","runs":70}

+++ Problem17.java
@@ -2,7 +2,7 @@ import acm.program.ConsoleProgram;
 
 public class Problem17 extends ConsoleProgram{
 	public void run() {
-		double a = readInt("Enter a: ");
+		double a = readDouble("Enter a: ");
 		
 	}
 }
